requirements.txt
django-watson
django-extensions
django_mail_admin
